Move sparkline to a different file
